+++ src/sdk/types/client.ts
@@ -93,6 +93,7 @@ export type RareClientContracts = {
   v4Quoter?: Address;
   cart?: Address;
   cartLens?: Address;
+  cartHashes?: Address;
 }
 
 export type SearchNamespace = {